% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{poisson_dat}
\alias{poisson_dat}
\title{Data to}
\format{
A list
\describe{
\item{X}{}
\item{S}{}
\item{Y}{}
\item{X.test}{}
}
}
\source{
{Created data to serve as an example.}
}
\usage{
poisson_dat
}
\description{
contains
}
\examples{
data(poisson_dat)
}
\keyword{datasets}
